requirements.txt
asn1crypto==0.24.0
bitrix24-python3-client==0.3.4
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
future==0.17.1
idna==2.8
multidimensional-urlencode==0.0.4
pycparser==2.19
python-dotenv==0.10.3
python-telegram-bot==11.1.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
